Add a wiki tables parser
