+++ libmozdata/release_calendar.py
@@ -2,81 +2,28 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from datetime import datetime
-import dateutil.parser
-
-try:
-    from HTMLParser import HTMLParser
-except:  # NOQA
-    from html.parser import HTMLParser
-import pytz
 import requests
+from . import utils
+from .wiki_parser import InvalidWiki, WikiParser
 
 
 CALENDAR_URL = 'https://wiki.mozilla.org/Release_Management/Calendar'
 _CALENDAR = None
 
 
-class InvalidWiki(Exception):
-    def __init__(self, s):
-        super(InvalidWiki, self).__init__(s)
+def _get_sub_versions(s):
+    s = s.split('.')
+    return list(map(int, s))
 
 
-class CalendarParser(HTMLParser):
-    def __init__(self):
-        HTMLParser.__init__(self)
-        self.rows = None
-        self.td = None
-        self.th = None
-
-    def handle_starttag(self, tag, attrs):
-        if self.rows is None and tag == 'table':
-            self.rows = []
-        if self.rows is not None:
-            if tag == 'tr':
-                self.rows.append([])
-            elif tag == 'td':
-                self.td = ''
-            elif tag == 'th':
-                self.th = ''
-
-    def handle_endtag(self, tag):
-        if self.rows is not None:
-            if tag == 'table':
-                raise StopIteration()
-            elif tag == 'td':
-                self.rows[-1].append(self.td)
-                self.td = None
-            elif tag == 'th':
-                self.rows[-1].append(self.th)
-                self.th = None
-
-    def handle_data(self, data):
-        data = data.replace('\\n', '\n')
-        data = data.strip()
-        if self.td is not None:
-            self.td = data
-        elif self.th is not None:
-            self.th = data
-
-    @staticmethod
-    def _get_date(s):
-        return pytz.utc.localize(dateutil.parser.parse(s))
-
-    @staticmethod
-    def _get_sub_versions(s):
-        s = s.split('.')
-        return list(map(int, s))
-
-    @staticmethod
-    def _get_versions(s):
-        fx = 'Firefox '
-        if not s.startswith(fx):
-            raise InvalidWiki('Invalid version format, expect: \"Firefox ...\"')
-        version = s[len(fx):]
-        versions = version.split(';')
-        versions = list(map(CalendarParser._get_sub_versions, versions))
-        return versions
+def _get_versions(s):
+    fx = 'Firefox '
+    if not s.startswith(fx):
+        raise InvalidWiki('Invalid version format, expect: \"Firefox ...\"')
+    version = s[len(fx):]
+    versions = version.split(';')
+    versions = list(map(_get_sub_versions, versions))
+    return versions
 
 
 def get_calendar():
@@ -85,10 +32,11 @@ def get_calendar():
         return _CALENDAR
 
     html = str(requests.get(CALENDAR_URL).text.encode('utf-8'))
-    parser = CalendarParser()
+    parser = WikiParser(tables=[0])
     try:
         parser.feed(html)
     except StopIteration:
+        table = parser.get_tables()[0]
         if [
             'Quarter',
             'Soft Freeze',
@@ -98,22 +46,22 @@ def get_calendar():
             'Release Date',
             'Release',
             'ESR',
-        ] != parser.rows[0]:
+        ] != table[0]:
             raise InvalidWiki('Column headers are wrong')
 
         _CALENDAR = []
-        for row in parser.rows[1:]:
+        for row in table[1:]:
             if row[0].startswith('Q'):
                 row = row[1:]
             _CALENDAR.append(
                 {
-                    'soft freeze': CalendarParser._get_date(row[0]),
-                    'merge': CalendarParser._get_date(row[1]),
-                    'central': CalendarParser._get_versions(row[2])[0][0],
-                    'beta': CalendarParser._get_versions(row[3])[0][0],
-                    'release date': CalendarParser._get_date(row[4]),
-                    'release': CalendarParser._get_versions(row[5])[0][0],
-                    'esr': CalendarParser._get_versions(row[6]),
+                    'soft freeze': utils.get_date_ymd(row[0]),
+                    'merge': utils.get_date_ymd(row[1]),
+                    'central': _get_versions(row[2])[0][0],
+                    'beta': _get_versions(row[3])[0][0],
+                    'release date': utils.get_date_ymd(row[4]),
+                    'release': _get_versions(row[5])[0][0],
+                    'esr': _get_versions(row[6]),
                 }
             )
         return _CALENDAR
@@ -121,7 +69,7 @@ def get_calendar():
 
 def get_next_release_date():
     cal = get_calendar()
-    now = pytz.utc.localize(datetime.utcnow())
+    now = utils.get_date_ymd('today')
     for c in cal:
         if now < c['release date']:
             return c['release date']
